run binary directly instead of passing by index.js
